// src/app/services/models/paged-transaction-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TransactionResponse } from '../models/transaction-response';
export interface PagedTransactionResponse {
  content?: Array<TransactionResponse>;
  first?: boolean;
  last?: boolean;
  pageNumber?: number;
  pageSize?: number;
  totalElements?: number;
  totalPages?: number;
}